// src/Sound/Sound.strings.d.ts
export declare const strings: {
    CTOR_AUDIO_ELEMENT_INVALID: string;
    CTOR_BUFFER_INVALID: string;
    CTOR_GET_MANAGER_VOLUME_INVALID: string;
    CTOR_OPTIONS_INVALID: string;
    GET_DURATION_AUDIO_ELEMENT_INVALID: string;
    GET_DURATION_BUFFER_INVALID: string;
    GET_FADE_GAIN_NODE_NODE_INVALID: string;
    GET_LOOP_AUDIO_ELEMENT_INVALID: string;
    GET_SOURCE_NODE_NODE_INVALID: string;
    GET_TRACK_POSITION_AUDIO_ELEMENT_INVALID: string;
    INITIALIZE_SOUND_FOR_WEB_AUDIO_BUFFER_INVALID: string;
    SET_LOOP_AUDIO_ELEMENT_INVALID: string;
    SET_TRACK_POSITION_AUDIO_ELEMENT_INVALID: string;
};